requirements.txt
numpy==1.20.1
opencv_python==4.5.5.64
Pillow==9.1.0
